requirements.txt
agno
google-genai
python-dotenv
streamlit
yfinance
beautifulsoup4
requests 
agno